==> 04 Disaggregation solution.ipynb <==
import numpy as np
import pandas as pd
import geopandas as gpd
import matplotlib.pyplot as plt
import shapely.geometry as sgeo
%matplotlib inline

# Disaggregation

The goal of this exercise is to learn how to switch from a zone-based and flow-based simulation to a trip-based setting. The idea is to take the flow matrix generated in the previous exercise and to generate individual trips.

## Disaggregating the flows

Instead of generating the true number of trips, we will only generate a small share to keep the calculations feasible. 

First, we read the generated flow data and the municipality shapes from the previous exercises:

df_flow = pd.read_parquet("data/flow.parquet")
df_municipalities = gpd.read_parquet("data/municipalities.parquet")

**Task**: Filter both the municipalities data frame and the flow data frame for the area of Paris (department 75). For the flow data set, make sure that only flows are kept that start within Paris and also end within Paris.

### Insert your code here

### SOLUTION START
df_flow = df_flow.reset_index()
df_flow = df_flow[df_flow["origin_id"].str.startswith("75") & df_flow["destination_id"].str.startswith("75")]

df_municipalities = df_municipalities[df_municipalities["municipality_id"].str.startswith("75")]

**Task**: Now convert the `weight` column (reference flows) or the `model` column (your choice) into a probability by dividing by the total:

### Insert your code here
# df_flow["probability"] = 

### SOLUTION START
df_flow["probability"] = df_flow["weight"] / df_flow["weight"].sum()

**Task**: Next, sample a trip table from the flow data frame, using *pandas*' `sample` method of the data frame, according to the `probability` of each relation. Make sure to pass `replace = True` so you can sample individual entries multiple times. Sample *1,000* trips.

### Insert your code here
# df_trips = 

### SOLUTION START
df_trips = df_flow.sample(1000, weights = "probability", replace = True)

Next, we clean up the trips data frame:

df_trips = df_trips[["origin_id", "destination_id"]]
df_trips.head()

## Sampling a departure time

Our next goal is to generate a departure time for each trip. We will only generate trips for the morning peak.

**Task**: Add a departure time to each trip by sampling from the following normal distribution:

$$
t \sim \mathcal{N}(\mu = 8.5, \sigma = 1)
$$

Save the departure time in seconds (seconds after midnight). Set any negative values to zero to avoid computational issues:

### Insert your code here
# df_trips["departure_time"] = 

### SOLUTION START
np.random.seed(0)
df_trips["departure_time"] = np.maximum(0.0, (8.5 + np.random.normal(size = len(df_trips))) * 3600)

**Task**: To make sure, plot a histogram of your generated departure times.

### Insert your code here

### SOLUTION START
plt.hist(df_trips["departure_time"] / 3600)

## Sampling origins and destinations

Next, we want to generate origin and destination points for the trips and show them on a map.

**Task**: We will follow a process that is not the most efficient, but straight-forward to follow. First, merge the municipality data frame onto the trips data frame such that a new column `origin_geometry` is created:

pd.DataFrame({ "origin_id": [], "destination_id": [], "departure_time": [], "origin_geometry": [] })

### Insert your code here
# df_trips = pd.merge(df_trips, df_municipalities.rename( # complete # ), on = "origin_id")

### SOLUTION START
df_trips = pd.merge(df_trips, df_municipalities.rename(columns = {
    "municipality_id": "origin_id", "geometry": "origin_geometry"
}), on = "origin_id")

**Task**: Now repeat the same to generate a new column `destination_geometry` for each trip:

### Insert your code here

### SOLUTION START
df_trips = pd.merge(df_trips, df_municipalities.rename(columns = {
    "municipality_id": "destination_id", "geometry": "destination_geometry"
}), on = "destination_id")

Let's clean the data set, we should have the following columns:

df_trips = df_trips[["origin_id", "destination_id", "departure_time", "origin_geometry", "destination_geometry"]]
assert len(df_trips) == 1000

*Geopandas* provides a useful method called `sample_points`, but it only acts on the active *geometry* column of a `GeoDataFrame`. First, we need to convert `df_trips` into a `GeoDataFrame` with `origin_geometry` as the active geometry column:

df_trips = gpd.GeoDataFrame(df_trips, geometry = "origin_geometry", crs = df_municipalities.crs)

Try the following code, it will take the polygon geometry of every trip's origin zone and sample a point from within that zone:

df_trips.sample_points(1)

**Task**: Override the (polygon) `origin_geometry` column in your trip table now with a sampled point from the respective zone:

### Insert your code here
# df_trips["origin_geometry"] = # ...

### SOLUTION START
df_trips["origin_geometry"] = df_trips.sample_points(1)

**Task**: Now, do the same with the destination. First, set the active geometry column of the data frame to `destination_geometry` (see `GeoDataFrame.set_geometry` and then override this column with a randomly sampled point:

### Insert your code here
# df_trips["origin_geometry"] = ...

### SOLUTION START
df_trips = df_trips.set_geometry("destination_geometry")
df_trips["destination_geometry"] = df_trips.sample_points(1)

Have a look at your data frame. You should see that both `origin_geometry` and `destination_geometry` are of type `POINT`. If they are *exactly* the same, you did something wrong in the previous cells:

df_trips.head()

We will now create a new geometry column which, instead of a `POINT` contains a `LINESTRING`, i.e. a connected line between `N` (in our case 2) points. For that, we make use of the `shapely.geometry` package, which we have imported as `sgeo` (see first cell).

df_trips["geometry"] = [
    sgeo.LineString(od) 
    for od in zip(df_trips["origin_geometry"], df_trips["destination_geometry"])
]

Try to understand the previous cell. What does it do?

**Task**: Set the *active geometry column* of the data frame to `geometry` and plot the data frame. What do you see?

### Insert your code here

### SOLUTION START
df_trips = df_trips.set_geometry("geometry")
df_trips.plot()

**Task**: In this notebook or using QGIS, plot the arrondissements of Paris together with your generated flows.

### Insert your code here or use QGIS

Let's save the generated trips for the next exercise:

df_trips.to_parquet("data/trips.parquet")

**Congratulations!** You should now be able to disaggregate a flow matrix for your course project (Exercise 3.1).
